+++ cmd/document-design-gateway/main.go
@@ -1,26 +1,14 @@
 package main
 
 import (
-	"context"
-	"net"
-	"net/http"
 	"os"
-	"os/signal"
-	"syscall"
 
-	"github.com/grpc-ecosystem/go-grpc-middleware/v2/interceptors/recovery"
 	"github.com/kinneko-de/restaurant-document-design-gateway/internal/app/document"
 	"github.com/kinneko-de/restaurant-document-design-gateway/internal/app/github/oauth"
 	"github.com/kinneko-de/restaurant-document-design-gateway/internal/app/operation/health"
 	"github.com/kinneko-de/restaurant-document-design-gateway/internal/app/operation/logger"
-	"github.com/kinneko-de/restaurant-document-design-gateway/internal/app/router"
+	"github.com/kinneko-de/restaurant-document-design-gateway/internal/app/server"
 	"github.com/rs/zerolog"
-	"github.com/rs/zerolog/log"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/codes"
-	googleHealth "google.golang.org/grpc/health"
-	"google.golang.org/grpc/health/grpc_health_v1"
-	"google.golang.org/grpc/status"
 )
 
 func main() {
@@ -43,8 +31,8 @@ func main() {
 	httpServerStopped := make(chan struct{})
 	grpcServerStarted := make(chan struct{})
 	grpcServerStopped := make(chan struct{})
-	go startHttpServer(httpServerStarted, httpServerStopped, ":8080")
-	go startGrpcServer(grpcServerStarted, grpcServerStopped, ":3110")
+	go server.StartHttpServer(httpServerStarted, httpServerStopped, ":8080")
+	go server.StartGrpcServer(grpcServerStarted, grpcServerStopped, ":3110")
 
 	<-grpcServerStarted
 	<-httpServerStarted
@@ -54,84 +42,4 @@ func main() {
 	<-httpServerStopped
 	logger.Logger.Info().Msg("Application stopped.")
 	os.Exit(0)
-}
-
-func startHttpServer(httpServerStarted chan struct{}, httpServerStopped chan struct{}, port string) {
-	router := router.SetupRouter()
-	var gracefulStop = make(chan os.Signal, 1)
-	signal.Notify(gracefulStop, syscall.SIGTERM, syscall.SIGINT)
-	logger.Logger.Debug().Msg("starting http server")
-
-	server := &http.Server{Addr: port, Handler: router}
-	go func() {
-		if err := server.ListenAndServe(); err != nil && err != http.ErrServerClosed {
-			logger.Logger.Error().Err(err).Msg("Failed to start http server")
-			os.Exit(50)
-		}
-	}()
-	close(httpServerStarted)
-
-	stop := <-gracefulStop
-	if err := server.Shutdown(context.Background()); err != nil {
-		logger.Logger.Error().Err(err).Msg("Failed to shutdown http server")
-	}
-	logger.Logger.Debug().Msgf("http server stopped. Received signal %s", stop)
-	close(httpServerStopped)
-}
-
-func startGrpcServer(grpcServerStarted chan struct{}, grpcServerStopped chan struct{}, port string) {
-	listener, err := net.Listen("tcp", port)
-	if err != nil {
-		logger.Logger.Error().Err(err).Msgf("Failed to listen on port %v", port)
-		os.Exit(51)
-	}
-
-	grpcServer := configureGrpcServer()
-	healthServer := googleHealth.NewServer()
-	grpc_health_v1.RegisterHealthServer(grpcServer, healthServer)
-	health.Initialize(healthServer)
-
-	var gracefulStop = make(chan os.Signal, 1)
-	signal.Notify(gracefulStop, syscall.SIGTERM, syscall.SIGINT)
-	logger.Logger.Debug().Msg("starting grpc server")
-
-	go func() {
-		if err := grpcServer.Serve(listener); err != nil {
-			logger.Logger.Error().Err(err).Msg("failed to serve grpc server")
-			os.Exit(52)
-		}
-	}()
-	close(grpcServerStarted)
-
-	stop := <-gracefulStop
-	healthServer.Shutdown()
-	grpcServer.GracefulStop()
-	logger.Logger.Debug().Msgf("http server stopped. received signal %s", stop)
-	close(grpcServerStopped)
-}
-
-func configureGrpcServer() *grpc.Server {
-	// Handling of panic to prevent crash from example nil pointer exceptions
-	logPanic := func(p any) (err error) {
-		log.Error().Any("method", p).Err(err).Msg("Recovered from panic.")
-		return status.Errorf(codes.Internal, "Internal server error occured.")
-	}
-
-	opts := []recovery.Option{
-		recovery.WithRecoveryHandler(logPanic),
-	}
-
-	grpcServer := grpc.NewServer(
-		grpc.UnaryInterceptor(
-			recovery.UnaryServerInterceptor(opts...),
-		),
-		grpc.StreamInterceptor(
-			recovery.StreamServerInterceptor(opts...),
-		),
-	)
-	RegisterAllGrpcServices(grpcServer)
-	return grpcServer
-}
-
-func RegisterAllGrpcServices(grpcServer *grpc.Server) {
 }
